+++ android/snapshot-helper/README.md
@@ -46,6 +46,51 @@ can run.
 `waitForIdleTimeoutMs` defaults to `500`, which is a maximum wait, not a fixed sleep. Direct helper
 invocations can pass `0` when immediate capture during ongoing animation is preferred.
 
+## One-Shot Modes
+
+Passing `-e mode snapshot|viewport|gesture` selects what a single instrumentation run does; `snapshot`
+is the default and matches the `Run` section above.
+
+```sh
+# Read the interactive-window viewport without capturing a snapshot.
+adb shell am instrument -w -e mode viewport \
+  com.callstack.agentdevice.snapshothelper/.SnapshotInstrumentation
+
+# Inject a planned touch gesture, described by a base64 JSON payload.
+adb shell am instrument -w -e mode gesture -e payloadBase64 "$PAYLOAD" \
+  com.callstack.agentdevice.snapshothelper/.SnapshotInstrumentation
+```
+
+The `gesture` payload is a base64-encoded JSON object using protocol `android-touch-plan-v1`:
+
+- `kind`: `swipe` (one pointer) or `transform` (two pointers, e.g. pinch/rotate)
+- `durationMs`: integer, `0`-`120000`
+- `pointers`: array of `{ pointerId, samples }`, `pointerId`s ordered from `0`, one pointer for
+  `swipe` and exactly two for `transform`
+- each pointer's `samples` is `{ offsetMs, x, y }[]` with at least two entries; `offsetMs` values
+  must be strictly increasing (equal offsets are only allowed when `durationMs` is `0`), the first
+  sample's `offsetMs` must be `0`, and the last must equal `durationMs`. For `transform` gestures,
+  both pointers must share the same `offsetMs` sequence.
+
+## Persistent Session
+
+Passing `-e sessionPort <port>` keeps the instrumentation alive after startup and serves repeated
+commands over a local TCP socket on `127.0.0.1:<port>`, instead of exiting after one snapshot. This
+avoids paying UiAutomation connect/teardown cost per call. Each line sent to the socket is one
+command; the instrumentation replies once per command and keeps the connection open for the next
+one:
+
+- `snapshot <requestId>` — capture and return an XML snapshot, same semantics as the default mode
+- `viewport <requestId>` — return interactive-window viewport bounds
+- `gesture <requestId> <payloadBase64>` — inject a planned touch gesture (same payload as the
+  one-shot `gesture` mode)
+- `quit <requestId>` — acknowledge and stop the session
+
+`viewport` and `gesture` responses are headers-only (no body): `kind`, `injectedEvents`,
+`elapsedMs` for `gesture`, and `x`, `y`, `width`, `height` for `viewport`. `snapshot` responses carry
+the XML body after the header block, as described below. The response protocol literal is always
+`android-snapshot-helper-v1`, regardless of session or one-shot transport.
+
 ## Output Contract
 
 The APK emits instrumentation status records using
@@ -64,10 +109,10 @@ Each XML chunk is sent with:
 - `chunkCount`
 - `payloadBase64`
 
-The final instrumentation result includes:
+The final instrumentation result for the default `snapshot` mode includes:
 
 - `ok=true`
-- `helperApiVersion=1`
+- `helperApiVersion=2`
 - `waitForIdleTimeoutMs`
 - `waitForIdleQuietMs`
 - `timeoutMs`
@@ -79,6 +124,10 @@ The final instrumentation result includes:
 - `nodeCount`
 - `truncated`
 - `elapsedMs`
+
+`viewport` and `gesture` one-shot results carry `agentDeviceProtocol`/`helperApiVersion`/
+`outputFormat` plus the mode-specific fields described under "One-Shot Modes" above, instead of the
+`snapshot`-mode fields listed here.
 
 Failures return `ok=false`, `errorType`, and `message` in the final result.
